DigitalOcean Spaces support

+++ app/Http/Controllers/RoomUploadMediaController.php
@@ -3,14 +3,15 @@
 namespace App\Http\Controllers;
 
 use App\ChunkUpload\UploadHandler;
+use App\Jobs\AddRoomMediaFile;
+use App\Models\QueueRoomFile;
 use App\Models\Room;
+use Closure;
 use Illuminate\Http\Request;
-use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rules\File;
 use Illuminate\Validation\ValidationException;
 use Pion\Laravel\ChunkUpload\Receiver\FileReceiver;
-use Spatie\MediaLibrary\MediaCollections\Exceptions\FileIsTooBig;
-use Spatie\MediaLibrary\Support\File;
-use Throwable;
 
 class RoomUploadMediaController extends Controller
 {
@@ -34,31 +35,50 @@ class RoomUploadMediaController extends Controller
             $file = $fileReceived->getFile();
 
             $extension = $file->getClientOriginalExtension();
-            $name = str_replace('.'.$extension, '', $file->getClientOriginalName());
-            $fileName = Str::random(20).'.'.$extension;
+            $fileName = str_replace('.'.$extension, '', $file->getClientOriginalName());
 
-            try {
-                $room
-                    ->addMedia($file)
-                    ->usingName($name)
-                    ->usingFileName($fileName)
-                    ->toMediaCollection();
-            } catch (FileIsTooBig $e) {
-                $fileSize = File::getHumanReadableSize(filesize($file->getPath().'/'.$file->getFilename()));
-                $maxFileSize = File::getHumanReadableSize(config('media-library.max_file_size'));
+            $validator = Validator::make(['file' => $file], [
+                'file' => [
+                    'required',
+                    'file',
+                    File::types(['mp4', 'mp3'])
+                        ->max(config('media-library.max_file_size')),
+                    function (string $attribute, mixed $value, Closure $fail) use ($fileName) {
+                        if (strlen($fileName) > 50) {
+                            $fail('檔名最多 50 個字');
+                        }
+                    },
+                ],
+            ]);
 
-                unlink($file->getPathname());
+            if ($validator->fails()) {
+                if (is_file($file->getPathname())) {
+                    unlink($file->getPathname());
+                }
 
-                throw ValidationException::withMessages([
-                    'file' => '檔案大小為 '.$fileSize.' 超過上限大小 '.$maxFileSize,
-                ]);
-            } catch (Throwable $e) {
-                unlink($file->getPathname());
+                // throw validate exception
+                $validator->validate();
 
-                throw $e;
+                return;
             }
 
-            return ['success' => '檔案上傳成功'];
+            $queueFile = QueueRoomFile::make();
+            $queueFile->name = $fileName;
+            $queueFile->path = $file->store('medias', ['disk' => 'local']);
+            $queueFile->disk = 'local';
+            $queueFile->save();
+
+            AddRoomMediaFile::dispatch($room, $queueFile);
+
+            if (is_file($file->getPathname())) {
+                unlink($file->getPathname());
+            }
+
+            if (config('queue.default') !== 'sync') {
+                return ['success' => '檔案上傳成功，等待處理媒體檔案...'];
+            }
+
+            return ['success' => null];
         }
 
         /** @var \App\ChunkUpload\UploadHandler */
